- Friends
- Voice/Video calls
- Mentions/Notifications
- Attachments
- DM channels

+++ internal/database/entities/avatar/schema.go
@@ -8,17 +8,19 @@ import (
 )
 
 const (
-	createAvatar       = `INSERT INTO gochat.avatars (id, user_id, object) VALUES (?, ?, ?)`
-	removeAvatar       = `DELETE FROM gochat.avatars WHERE id = ?`
-	getAvatar          = `SELECT id, user_id, object FROM gochat.avatars WHERE id = ?`
-	getAvatarsByUserId = `SELECT id, user_id, object FROM gochat.avatars WHERE user_id = ?`
+	// Placeholder row with TTL and done=false
+	createAvatar     = `INSERT INTO gochat.avatars (id, user_id, done, filesize) VALUES (?, ?, false, ?) USING TTL ?`
+	getAvatar        = `SELECT id, user_id, url, content_type, width, height, filesize, done FROM gochat.avatars WHERE user_id = ? AND id = ?`
+	doneAvatar       = `UPDATE gochat.avatars USING TTL 0 SET done = true, content_type = ?, url = ?, height = ?, width = ?, filesize = ? WHERE user_id = ? AND id = ?`
+	removeAvatar     = `DELETE FROM gochat.avatars WHERE user_id = ? AND id = ?`
+	getAvatarsByUser = `SELECT id, user_id, url, content_type, width, height, filesize, done FROM gochat.avatars WHERE user_id = ?`
 )
 
-func (e *Entity) CreateAvatar(ctx context.Context, id, userId int64, object string) error {
+func (e *Entity) CreateAvatar(ctx context.Context, id, userId, ttlSeconds, fileSize int64) error {
 	err := e.c.Session().
 		Query(createAvatar).
 		WithContext(ctx).
-		Bind(id, userId, object).
+		Bind(id, userId, fileSize, ttlSeconds).
 		Exec()
 	if err != nil {
 		return fmt.Errorf("unable to create avatar: %w", err)
@@ -26,11 +28,36 @@ func (e *Entity) CreateAvatar(ctx context.Context, id, userId int64, object stri
 	return nil
 }
 
-func (e *Entity) RemoveAvatar(ctx context.Context, id int64) error {
+func (e *Entity) GetAvatar(ctx context.Context, id, userId int64) (model.Avatar, error) {
+	var a model.Avatar
+	err := e.c.Session().
+		Query(getAvatar).
+		WithContext(ctx).
+		Bind(userId, id).
+		Scan(&a.Id, &a.UserId, &a.URL, &a.ContentType, &a.Width, &a.Height, &a.FileSize, &a.Done)
+	if err != nil {
+		return a, fmt.Errorf("unable to get avatar: %w", err)
+	}
+	return a, nil
+}
+
+func (e *Entity) DoneAvatar(ctx context.Context, id, userId int64, contentType, url *string, height, width, fileSize *int64) error {
+	err := e.c.Session().
+		Query(doneAvatar).
+		WithContext(ctx).
+		Bind(contentType, url, height, width, fileSize, userId, id).
+		Exec()
+	if err != nil {
+		return fmt.Errorf("unable to done avatar: %w", err)
+	}
+	return nil
+}
+
+func (e *Entity) RemoveAvatar(ctx context.Context, id, userId int64) error {
 	err := e.c.Session().
 		Query(removeAvatar).
 		WithContext(ctx).
-		Bind(id).
+		Bind(userId, id).
 		Exec()
 	if err != nil {
 		return fmt.Errorf("unable to remove avatar: %w", err)
@@ -38,32 +65,19 @@ func (e *Entity) RemoveAvatar(ctx context.Context, id int64) error {
 	return nil
 }
 
-func (e *Entity) GetAvatar(ctx context.Context, id int64) (model.Avatar, error) {
-	var a model.Avatar
-	err := e.c.Session().
-		Query(getAvatar).
-		WithContext(ctx).
-		Bind(id).
-		Exec()
-	if err != nil {
-		return a, fmt.Errorf("unable to get avatar: %w", err)
-	}
-	return a, nil
-}
-
+// GetAvatarsByUserId returns list of avatars for a user
 func (e *Entity) GetAvatarsByUserId(ctx context.Context, userId int64) ([]model.Avatar, error) {
 	var avatars []model.Avatar
 	iter := e.c.Session().
-		Query(getAvatarsByUserId).
+		Query(getAvatarsByUser).
 		WithContext(ctx).
 		Bind(userId).
 		Iter()
 	var a model.Avatar
-	for iter.Scan(&a.Id, &a.UserId, &a.Object) {
+	for iter.Scan(&a.Id, &a.UserId, &a.URL, &a.ContentType, &a.Width, &a.Height, &a.FileSize, &a.Done) {
 		avatars = append(avatars, a)
 	}
-	err := iter.Close()
-	if err != nil {
+	if err := iter.Close(); err != nil {
 		return nil, fmt.Errorf("unable to get avatars: %w", err)
 	}
 	return avatars, nil
